Updates todo controller
